עובדדדדדד המעקב אחרי המסלולללללל תודה לאלללללללדלדכלדגכלדגכדגכדגכדגכדג

+++ app/src/main/java/com/example/beta/CreatingPath.java
@@ -1,21 +1,22 @@
 package com.example.beta;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.graphics.Canvas;
 import android.graphics.Color;
+import android.graphics.Paint;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
-import android.os.Build;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -23,6 +24,7 @@ import android.view.WindowManager;
 import android.widget.ImageView;
 import android.widget.Toast;
 
+import com.example.beta.activities.AddingPlace;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.storage.StorageReference;
@@ -50,10 +52,12 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
     int x,y,z;
 
 
-    Bitmap bMap,savedmap;
+
+    Bitmap bMap,savedmap,bMap2;
 
     boolean startwalkingtrue=false;
     boolean usersPermission=false;
+    boolean downladpic=false;
     int steps=0;
 
     AlertDialog.Builder adb;
@@ -70,36 +74,58 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
         yS = gi.getIntExtra("y", -1);
         mapId = gi.getStringExtra("mapId");
 
+
         sensorManager=(SensorManager) getSystemService(Context.SENSOR_SERVICE);
-        //accelerometer=sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
-        //sensorManager.registerListener(CreatingPath.this,accelerometer,SensorManager.SENSOR_DELAY_NORMAL);
-        //magnometer=sensorManager.getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD);
-        //sensorManager.registerListener(CreatingPath.this,magnometer,SensorManager.SENSOR_DELAY_NORMAL);
+        accelerometer=sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
+        sensorManager.registerListener(CreatingPath.this,accelerometer,SensorManager.SENSOR_DELAY_NORMAL);
+        magnometer=sensorManager.getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD);
+        sensorManager.registerListener(CreatingPath.this,magnometer,SensorManager.SENSOR_DELAY_NORMAL);
 
         getWindow().addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
 
 
         //sensorManager=(SensorManager) getSystemService(SENSOR_SERVICE);
 
-        mStepCounter = sensorManager.getDefaultSensor(Sensor.TYPE_STEP_COUNTER);
-        if(mStepCounter!=null) {
-            Toast.makeText(this, "step counter found", Toast.LENGTH_SHORT);
-            sensorManager.registerListener(this, mStepCounter, SensorManager.SENSOR_DELAY_NORMAL);
+
+      /*  if(sensorManager.getDefaultSensor(Sensor.TYPE_STEP_COUNTER)!=null){
+            mStepCounter=sensorManager.getDefaultSensor(Sensor.TYPE_STEP_COUNTER);
+            isCounterSensorPresent=true;
         }
-        else
-        {
-            Toast.makeText(this, "No step counter!!!", Toast.LENGTH_SHORT);
+        else{
+            isCounterSensorPresent=false;
+        }
+       */
+        try {
+            mStepCounter = sensorManager.getDefaultSensor(Sensor.TYPE_STEP_COUNTER);
+            if(mStepCounter!=null) {
+                Toast.makeText(this, "step counter found", Toast.LENGTH_SHORT);
+                sensorManager.registerListener(this, mStepCounter, SensorManager.SENSOR_DELAY_NORMAL);
+            }
+            else
+            {
+                Toast.makeText(this, "No step counter!!!", Toast.LENGTH_SHORT);
+            }
+        }
+        catch(Exception ex){
+         Log.d("Debug","Exception:"+ ex.toString());
         }
 
 
+        final ProgressDialog pd=ProgressDialog.show(this,"PIC","Downloading map.... ",true);
         StorageReference refImages=refStor.child(mapId+".jpg");
-        final long MAX_SIZE = 1024*1024;
+        final long MAX_SIZE = 2500*2500;
         refImages.getBytes(MAX_SIZE).addOnSuccessListener(new OnSuccessListener<byte[]>() {
             @Override
             public void onSuccess(byte[] bytes) {
-                bMap = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
-                savedmap= BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
+                bMap2 = BitmapFactory.decodeByteArray(bytes, 0, bytes.length);
+                bMap=bMap2.copy(bMap2.getConfig(),true);
+                savedmap= bMap2.copy(bMap2.getConfig(),true);
                 mapIV.setImageBitmap(bMap);
+                downladpic=true;
+                startwalkingtrue=false;
+               // bMap=createImage(1000,1000,Color.BLACK);
+               // mapIV.setImageBitmap(bMap);
+                pd.dismiss();
             }
         }).addOnFailureListener(new OnFailureListener() {
             @Override
@@ -108,7 +134,7 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
             }
         });
 
-        startwalkingtrue=false;
+
 
 
 
@@ -116,123 +142,100 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
 
     @Override
     public void onSensorChanged(SensorEvent sensorEvent) {
-        //if (sensorEvent.sensor.getType() != Sensor.TYPE_STEP_COUNTER) {return;}
-//        public static final int TYPE_ORIENTATION = 3;
-//        public static final int TYPE_POSE_6DOF = 28;
-//        public static final int TYPE_PRESSURE = 6;
-//        public static final int TYPE_PROXIMITY = 8;
-//        public static final int TYPE_RELATIVE_HUMIDITY = 12;
-//        public static final int TYPE_ROTATION_VECTOR = 11;
-//        public static final int TYPE_SIGNIFICANT_MOTION = 17;
-//        public static final int TYPE_STATIONARY_DETECT = 29;
-//        public static final int TYPE_STEP_COUNTER = 19;
-//        public static final int TYPE_STEP_DETECTOR = 18;
+        if (downladpic) {
+            if (sensorEvent.sensor.getType() == Sensor.TYPE_STEP_COUNTER) {
+                if (startwalkingtrue) {
+                    if (sensorEvent.sensor == mStepCounter) {
+                        steps++;
+                        if (steps == 3) {
+                            steps = 0;
+                            if (rotation >= 22.5 && rotation <= 67.5) {
+                         //       if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())){
+                                for (int i = 0; i < 2; i++) {
+                                        xS++;
+                                        yS--;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                        //        }
+                                }
+                            } else if (rotation >= 112.5 && rotation <= 157.5) {
+                        //        if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 2; i++) {
+                                        xS++;
+                                        yS++;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                       //         }
+                            } else if (rotation >= 202.5 && rotation <= 247.5) {
+                       //         if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 2; i++) {
+                                        xS--;
+                                        yS++;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                          //      }
+                            } else if (rotation >= 292.5 && rotation <= 337.5) {
+                            //    if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 2; i++) {
+                                        xS--;
+                                        yS--;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                          //      }
+                            } else if (rotation > 67.5 && rotation < 112.5) {
+                        //        if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 3; i++) {
+                                        xS++;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                          //      }
+                            } else if (rotation > 157.5 && rotation < 202.5) {
+                        //        if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
+                                    for (int i = 0; i < 3; i++) {
+                                        yS++;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                          //      }
+                            } else if (rotation > 247.5 && rotation < 292.5) {
+                         //       if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 3; i++) {
+                                        xS--;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                          //      }
+                            } else if (rotation > 337.5 || rotation < 22.5) {
+                         //       if(dontexceed(xS,yS,bMap.getWidth(),bMap.getHeight())) {
+                                    for (int i = 0; i < 3; i++) {
+                                        yS--;
+                                        bMap.setPixel(xS, yS, Color.WHITE);
+                                    }
+                        //        }
+                            }
+                            mapIV.setImageBitmap(bMap);
+                        }
+                    }
+                }
+            } else if (sensorEvent.sensor.getType() == Sensor.TYPE_ACCELEROMETER)
+                mGravity = sensorEvent.values;
 
-        if (sensorEvent.sensor.getType() == Sensor.TYPE_STEP_COUNTER) {
-            Log.d("myapp1", "Sensor is: STEP COUNTER");
-            Toast.makeText(this, "step counter changed", Toast.LENGTH_SHORT);
-        }
-        else {
-            Log.d("myapp1", "Another Sensor: " + Sensor.TYPE_STEP_COUNTER);
-            Toast.makeText(this, "another sensor changed", Toast.LENGTH_SHORT);
-        }
+            else if (sensorEvent.sensor.getType() == Sensor.TYPE_MAGNETIC_FIELD)
+                mGeomagnetic = sensorEvent.values;
 
-        return;
+            if (mGravity != null && mGeomagnetic != null) {
+                float R[] = new float[9];
+                float I[] = new float[9];
 
-//        if(startwalkingtrue) {
-//                if (sensorEvent.sensor == mStepCounter) {
-//                        steps++;
-//                        if (steps == 3) {
-//                            steps = 0;
-//                            if (rotation >= 22.5 && rotation <= 67.5) {
-//                                 for (int i = 0; i < 2; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS++;
-//                                        yS--;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation >= 112.5 && rotation <= 157.5) {
-//                                for (int i = 0; i < 2; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS++;
-//                                        yS++;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation >= 202.5 && rotation <= 247.5) {
-//                                for (int i = 0; i < 2; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS--;
-//                                        yS++;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation >= 292.5 && rotation <= 337.5) {
-//                                for (int i = 0; i < 2; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS--;
-//                                        yS--;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation > 67.5 && rotation < 112.5) {
-//                                for (int i = 0; i < 3; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS++;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation > 157.5 && rotation < 202.5) {
-//                                for (int i = 0; i < 3; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        yS++;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation > 247.5 && rotation < 292.5) {
-//                                for (int i = 0; i < 3; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        xS--;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            } else if (rotation > 337.5 || rotation < 22.5) {
-//                                for (int i = 0; i < 3; i++) {
-//                                    if (dontexceed(xS, yS, bMap.getWidth(), bMap.getHeight())) {
-//                                        yS--;
-//                                        bMap.setPixel(xS, yS, Color.WHITE);
-//                                    }
-//                                }
-//                            }
-//                             mapIV.setImageBitmap(bMap);
-//                        }
-//                    }
-//                }
+                if (SensorManager.getRotationMatrix(R, I, mGravity, mGeomagnetic)) {
 
-        }
-    /*    if (sensorEvent.sensor.getType() == Sensor.TYPE_ACCELEROMETER)
-            mGravity = sensorEvent.values;
+                    // orientation contains azimut, pitch and roll
+                    float orientation[] = new float[3];
+                    SensorManager.getOrientation(R, orientation);
 
-        if (sensorEvent.sensor.getType() == Sensor.TYPE_MAGNETIC_FIELD)
-            mGeomagnetic = sensorEvent.values;
-
-        if (mGravity != null && mGeomagnetic != null) {
-            float R[] = new float[9];
-            float I[] = new float[9];
-
-            if (SensorManager.getRotationMatrix(R, I, mGravity, mGeomagnetic)) {
-
-                // orientation contains azimut, pitch and roll
-                float orientation[] = new float[3];
-                SensorManager.getOrientation(R, orientation);
-
-                azimuth = orientation[0];
-                rotation = (float) (( Math.toDegrees( azimuth ) + 360 ) % 360);
+                    azimuth = orientation[0];
+                    rotation = (float) ((Math.toDegrees(azimuth) + 360) % 360);
+                }
             }
-        }*/
-
-
+        }
+    }
 
     @Override
     public void onAccuracyChanged(Sensor sensor, int i) {
@@ -275,40 +278,46 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
 
     public void cancel(View view) {
         startwalkingtrue=false;
+        usersPermission=false;
         bMap=savedmap.copy(savedmap.getConfig(),true);
         mapIV.setImageBitmap(bMap);
     }
 
     public void pin(View view) {
-        StorageReference refImages=refStor.child(mapId+".jpg");
-        refImages.delete();
+        if(startwalkingtrue) {
+            StorageReference refImages = refStor.child(mapId + ".jpg");
+            refImages.delete();
 
-        ByteArrayOutputStream baos=new ByteArrayOutputStream();
-        bMap.compress(Bitmap.CompressFormat.JPEG,100,baos);
-        byte[] data=baos.toByteArray();
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            bMap.compress(Bitmap.CompressFormat.JPEG, 100, baos);
+            byte[] data = baos.toByteArray();
 
-        /**
-         * uploading the compressed bitmap
-         */
-        UploadTask uploadTask =refImages.putBytes(data);
-        uploadTask.addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-                Toast.makeText(CreatingPath.this, "There was a problem", Toast.LENGTH_SHORT).show();
-            }
-        }).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-            @Override
-            public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
+            /**
+             * uploading the compressed bitmap
+             */
+            UploadTask uploadTask = refImages.putBytes(data);
+            uploadTask.addOnFailureListener(new OnFailureListener() {
+                @Override
+                public void onFailure(@NonNull Exception exception) {
+                    Toast.makeText(CreatingPath.this, "There was a problem", Toast.LENGTH_SHORT).show();
+                }
+            }).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
+                @Override
+                public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
 
-            }
-        });
+                }
+            });
 
-        Intent si = new Intent(CreatingPath.this,AddingPlace.class);
-        si.putExtra("mapId",mapId);
-        si.putExtra("x",xS);
-        si.putExtra("y",yS);
-        si.putExtra("firstTime",false);
-        startActivity(si);
+            Intent si = new Intent(CreatingPath.this, AddingPlace.class);
+            si.putExtra("mapId", mapId);
+            si.putExtra("x", xS);
+            si.putExtra("y", yS);
+            si.putExtra("firstTime", false);
+            startActivity(si);
+        }
+        else{
+            Toast.makeText(this, "start walking", Toast.LENGTH_SHORT);
+        }
     }
 
     @Override
@@ -330,5 +339,13 @@ public class CreatingPath extends AppCompatActivity implements SensorEventListen
     }
 
 
+    public static Bitmap createImage(int width, int height, int color) {
+        Bitmap bitmap = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888);
+        Canvas canvas = new Canvas(bitmap);
+        Paint paint = new Paint();
+        paint.setColor(color);
+        canvas.drawRect(0F, 0F, (float) width, (float) height, paint);
+        return bitmap;
+    }
 
 }
